Move helper methods out of controller
