+++ Web/Edubase.Web.UI/Controllers/AdminController.cs
@@ -74,8 +74,8 @@ namespace Edubase.Web.UI.Controllers
 
             var webLogMessages = await _webLogItemRepository.GetWithinDateRange(startDate.Value, endDate.Value);
 
-            webLogMessages = FilterByAllTextColumns(webLogMessages, queryString);
-            webLogMessages = FilterPurgeZeroLogsMessage(webLogMessages, includePurgeZeroLogsMessage);
+            webLogMessages = WebLogItemRepository.FilterByAllTextColumns(webLogMessages, queryString);
+            webLogMessages = WebLogItemRepository.FilterPurgeZeroLogsMessage(webLogMessages, includePurgeZeroLogsMessage);
 
             var viewModel = new LogsViewModel
             {
@@ -98,102 +98,5 @@ namespace Edubase.Web.UI.Controllers
             return DateTime.TryParse($"{year}-{month}-{day}", out _);
         }
 
-        private List<AZTLoggerMessages> FilterPurgeZeroLogsMessage(List<AZTLoggerMessages> webLogMessages, bool includePurgeZeroLogsMessage)
-        {
-            if (includePurgeZeroLogsMessage)
-            {
-                return webLogMessages;
-            }
-
-            webLogMessages = webLogMessages
-                .Where(m => !m.Message.Equals("LOG PURGE REPORT: There were 0 logs purged from storage."))
-                .ToList();
-
-            return webLogMessages;
-        }
-
-        private static List<AZTLoggerMessages> FilterByAllTextColumns(List<AZTLoggerMessages> webLogMessages,
-            string queryString)
-        {
-            if (string.IsNullOrWhiteSpace(queryString))
-            {
-                return webLogMessages;
-            }
-
-            var lowerQueryString = queryString.ToLowerInvariant();
-
-            webLogMessages = webLogMessages.Where(m =>
-                {
-                    if (!string.IsNullOrWhiteSpace(m.Id) && m.Id.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.Level) && m.Level.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.Environment) && m.Environment.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.Message) && m.Message.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.Exception) && m.Exception.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.Url) && m.Url.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.UserAgent) && m.UserAgent.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.ClientIpAddress) && m.ClientIpAddress.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.ReferrerUrl) && m.ReferrerUrl.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.HttpMethod) && m.HttpMethod.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.RequestJsonBody) && m.RequestJsonBody.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.UserId) && m.UserId.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(m.UserName) && m.UserName.ToLowerInvariant().Contains(lowerQueryString))
-                    {
-                        return true;
-                    }
-
-                    return false;
-                }
-            ).ToList();
-
-            return webLogMessages;
-        }
     }
 }
